STP-2 Fix sanitization of indices
Indices can have values between 0 and size of view (included). Value equal to size of the view is out of boundary indicator.
Special End value is equal to std::numeric_limits<int>() and is an easy way to indicate that any operation should be performed until boundary of a view is achieved. This does not require preliminary knowledge of a size of a container. Only drawback is that containers with size 2147483647 elements can not be handled correctly. However I think this is not an ussual case and it will be handled in the compile time to disallow of such usage (just to be safe)
Index can be sanitize even if rvalue reference is passed.
API has been changed as start/stop/step is more convenient way than start/step/stop (IMHO)